+++ changelog.md
@@ -3,6 +3,9 @@
 ## Upcoming
 * Removed Gravity Defying
   * Was too strong, and nerfing it to the ground would make it useless on its own and too strong with M.A. Jump
+* Added an actual Enchanted Cleansed Lapis item
+  * I have no idea why I wanted to get myself into trouble by using an "enchanted" cleansed lapis
+  * Also Ancient Lapis will now be transformed to Enchanted Cleansed Lapis
 * Refactored enchanting feature tooltip
   * It was becoming spaghetti and there were untranslatable strings
   * Also fixed Base Enchantability showing on non-enchantable items